Shift activitywithcount to activity package
