+++ src/main/java/seedu/planner/model/activity/ActivityWithCount.java
@@ -1,0 +1,38 @@
+package seedu.planner.model.activity;
+
+import java.util.Objects;
+
+/**
+ * Represents the number of times the Activity appears in the Timetable.
+ */
+public class ActivityWithCount implements Comparable<ActivityWithCount> {
+    private final Activity activity;
+    private final long count;
+
+    public ActivityWithCount(Activity activity, long count) {
+        this.activity = activity;
+        this.count = count;
+    }
+
+    public Activity getActivity() {
+        return activity;
+    }
+
+    @Override
+    public int compareTo(ActivityWithCount o) {
+        return (int) (count - o.count);
+    }
+
+    @Override
+    public boolean equals(Object other) {
+        return other == this
+                || (other instanceof ActivityWithCount
+                && activity.equals((((ActivityWithCount) other).activity))
+                && count == ((ActivityWithCount) other).count);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(activity, count);
+    }
+}
